requirements.txt
pytest
playwright
python-dotenv
pytest-html